pyproject.toml
[tool.poetry]
name = "dinosavi"
version = "0.1.0"
description = "Self-supervised learning of Video Object Segmentation using DINOSAUR and SAVi"
authors = ["John-Henry Lim <42513874+Interpause@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "dinosavi"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
torch = "^2.0.0"
torchvision = "^0.15.1"
av = "^10.0.0"
pillow = "^9.4.0"
numpy = "^1.24.2"
einops = "^0.6.0"
hydra-core = "^1.3.1"
omegaconf = "^2.3.0"
rich = "^13.3.1"
torchinfo = "^1.7.2"
tensorboard = "^2.12.0"
transformers = "^4.26.1"
xxhash = "^3.2.0"
diskcache = "^5.4.0"
scipy = "^1.10.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
pydocstyle = "^6.3.0"
mypy = "^1.0.1"
ipykernel = "^6.21.2"
ipywidgets = "^8.0.4"
matplotlib = "^3.7.1"
ipympl = "^0.9.3"

[tool.black]

[tool.isort]
profile = "black"

[tool.pydocstyle]
convention = "google"

[tool.mypy]
implicit_optional = true
allow_redefinition = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
